move font files out, fix missing globals

+++ serialization.go
@@ -20,8 +20,8 @@ import (
 func SerializeChunk(c Chunk) []byte {
 	w := uint16(c.Bounds.W())
 	h := uint16(c.Bounds.H())
-	x := uint16(c.Bounds.Min.X / ChunkSize)
-	y := uint16(c.Bounds.Min.Y / ChunkSize)
+	x := uint16(c.Bounds.Min.X / float64(ChunkSize))
+	y := uint16(c.Bounds.Min.Y / float64(ChunkSize))
 
 	out := make([]byte, 4+(2*w*h)) // 4 bytes for x and y then 2 bytes for each tile
 	curVal := make([]byte, 2)      // Placeholder for a little-endian representation of the current value.
@@ -53,7 +53,7 @@ func DeserializeChunk(b []byte) *Chunk {
 	x := int16(binary.LittleEndian.Uint16(b[0:2]))
 	y := int16(binary.LittleEndian.Uint16(b[2:4]))
 
-	c := NewChunk(pixel.R(float64(x*ChunkSize), float64(y*ChunkSize), float64((x+1)*ChunkSize), float64(y+1)*ChunkSize))
+	c := NewChunk(pixel.R(float64(x)*float64(ChunkSize), float64(y)*float64(ChunkSize), float64(x+1)*float64(ChunkSize), float64(y+1)*float64(ChunkSize)))
 
 	byteIndex := 4
 	for x := range c.Tiles {
